+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Note: You may need to set "DVD order" in your settings again as the internal set
  - Movie Renamer: `<director>` placeholder is now supported
  - Movie: Add TMDb ID field to UI (#1022)
  - Movie: Add buttons that take you to the movie's IMDb/TMDb page (#684)
+ - Movie Filter: Add "Has TMDb ID"/"No TMDb ID" filters (#684)
 
 ### Internal Improvements and Changes